Read config file on options init. Fixes #22
Adds a new read_config method to the Options class that takes an IO
object and reads configuration options into itself. The method
defaults to reading from `.s3_sync` in the MM project root if no IO
is passed to it.

Adds a call to read_config in the Options constructor so the config
file will be read no matter where it is instantiated.