Add support for explicit specification of content-types for S3 paths/keys.
### What
Add support for explicit specification of content-types for S3 paths/keys.

### Why
The gem currently infers the content-type given the extension and there are many cases where the extension does not or cannot accurately reflect the content-type. An example of this is when the file has no extension but may be text/html, or the file may be a custom file type like `kml` where the content type is `text/xml`.

While this isn't as awesome as automatically pulling the content-type from the sitemap, the sitemap get's generated after_configuration and can be tricky and I think this will solve most of the problem cases I've seen people asking about online and can fill the gap until the next version of the gem is released using the new Middleman extension callbacks.

### Usage
During configuration a hash of S3 paths/keys to content-types can be defined. Because it's left up to the user to define there's a great deal of freedom here.

#### Example 1
A user could configure hardcoded paths to content types.
```ruby
activate :s3_sync do |s3_sync|
  ...
  s3_sync.content_types = {
    "sitemap.xml" => "text/xml",
    "data.json" => "application/json",
    "extensionlesspage" => "text/html"
  }
end
```

#### Example 2
If a user is generating dynamic pages they can keep track of the paths and set their content-types to be html.
```ruby
page_content_types = {}
10.times do |page_number|
  path = "page1.html"
  page(path, :proxy => "/page.html, :content_type => "text/html")
  page_content_types[path] = "text/html"
end

activate :s3_sync do |s3_sync|
  ...
  s3_sync.content_types = page_content_types
end
```